remove old parser, simplify int
